+++ wallet/network/errors.go
@@ -2,20 +2,6 @@ package network
 
 import "fmt"
 
-type AlreadyExistsError struct {
-	Name string
-}
-
-func NewAlreadyExistsError(n string) AlreadyExistsError {
-	return AlreadyExistsError{
-		Name: n,
-	}
-}
-
-func (e AlreadyExistsError) Error() string {
-	return fmt.Sprintf("network \"%s\" already exists", e.Name)
-}
-
 type DoesNotExistError struct {
 	Name string
 }