+++ wallet/service/store/v1/store_test.go
@@ -1,17 +1,19 @@
 package v1_test
 
 import (
-	"os"
-	"path/filepath"
 	"testing"
+	"time"
 
+	vgencoding "code.vegaprotocol.io/vega/libs/encoding"
 	vgrand "code.vegaprotocol.io/vega/libs/rand"
 	vgtest "code.vegaprotocol.io/vega/libs/test"
 	"code.vegaprotocol.io/vega/paths"
+	"code.vegaprotocol.io/vega/wallet/service"
 	storeV1 "code.vegaprotocol.io/vega/wallet/service/store/v1"
 	v1 "code.vegaprotocol.io/vega/wallet/service/v1"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"go.uber.org/zap"
 )
 
 func TestFileStoreV1(t *testing.T) {
@@ -22,6 +24,9 @@ func TestFileStoreV1(t *testing.T) {
 	t.Run("Verifying existing RSA keys succeeds", testFileStoreV1VerifyingExistingRSAKeysSucceeds)
 	t.Run("Getting non-existing RSA keys fails", testFileStoreV1GetNonExistingRSAKeysFails)
 	t.Run("Getting existing RSA keys succeeds", testFileStoreV1GetExistingRSAKeysSucceeds)
+	t.Run("Getting config while not being initialised succeeds", testFileStoreV1GetConfigWhileNotInitialisedSucceeds)
+	t.Run("Saving config succeeds", testFileStoreV1SavingConfigSucceeds)
+	t.Run("Verifying config succeeds", testFileStoreV1VerifyingConfigSucceeds)
 }
 
 func testNewStoreSucceeds(t *testing.T) {
@@ -163,6 +168,88 @@ func testFileStoreV1GetExistingRSAKeysSucceeds(t *testing.T) {
 	assert.Equal(t, keys, returnedKeys)
 }
 
+func testFileStoreV1GetConfigWhileNotInitialisedSucceeds(t *testing.T) {
+	vegaHome := newVegaHome(t)
+
+	// given
+	s := initialiseFromPath(t, vegaHome)
+
+	// when
+	cfg, err := s.GetConfig()
+
+	// then
+	require.NoError(t, err)
+	assert.Equal(t, service.DefaultConfig(), cfg)
+}
+
+func testFileStoreV1SavingConfigSucceeds(t *testing.T) {
+	vegaHome := newVegaHome(t)
+
+	// given
+	s := initialiseFromPath(t, vegaHome)
+	originalCfg := &service.Config{
+		LogLevel: vgencoding.LogLevel{
+			Level: zap.DebugLevel,
+		},
+		Server: service.ServerConfig{
+			Port: 123456789,
+			Host: vgrand.RandomStr(5),
+		},
+		APIV1: service.APIV1Config{
+			MaximumTokenDuration: vgencoding.Duration{
+				Duration: 234 * time.Second,
+			},
+		},
+	}
+
+	// when
+	err := s.SaveConfig(originalCfg)
+
+	// then
+	require.NoError(t, err)
+
+	// when
+	cfg, err := s.GetConfig()
+
+	// then
+	require.NoError(t, err)
+	assert.Equal(t, originalCfg, cfg)
+}
+
+func testFileStoreV1VerifyingConfigSucceeds(t *testing.T) {
+	vegaHome := newVegaHome(t)
+
+	// given
+	s := initialiseFromPath(t, vegaHome)
+	originalCfg := &service.Config{
+		LogLevel: vgencoding.LogLevel{
+			Level: zap.DebugLevel,
+		},
+		Server: service.ServerConfig{
+			Port: 123456789,
+			Host: vgrand.RandomStr(5),
+		},
+		APIV1: service.APIV1Config{
+			MaximumTokenDuration: vgencoding.Duration{
+				Duration: 234 * time.Second,
+			},
+		},
+	}
+
+	// when
+	err := s.SaveConfig(originalCfg)
+
+	// then
+	require.NoError(t, err)
+
+	// when
+	cfg, err := s.GetConfig()
+
+	// then
+	require.NoError(t, err)
+	assert.Equal(t, originalCfg, cfg)
+}
+
 func initialiseFromPath(t *testing.T, vegaHome *paths.CustomPaths) *storeV1.Store {
 	t.Helper()
 	s, err := storeV1.InitialiseStore(vegaHome)
@@ -175,14 +262,7 @@ func initialiseFromPath(t *testing.T, vegaHome *paths.CustomPaths) *storeV1.Stor
 
 func newVegaHome(t *testing.T) *paths.CustomPaths {
 	t.Helper()
-	rootPath := filepath.Join("/tmp", "vegawallet", vgrand.RandomStr(10))
-	t.Cleanup(func() {
-		if err := os.RemoveAll(rootPath); err != nil {
-			t.Fatalf("couldn't remove vega home: %v", err)
-		}
-	})
-
-	return &paths.CustomPaths{CustomHome: rootPath}
+	return &paths.CustomPaths{CustomHome: t.TempDir()}
 }
 
 func rsaKeysHome(t *testing.T, vegaHome *paths.CustomPaths) string {